Squash some printf debug log warnings

+++ src/hardware/pcspeaker.cpp
@@ -303,7 +303,7 @@ void PCSPEAKER_SetCounter(Bitu cntr, Bitu mode)
 		break;
 	default:
 #if C_DEBUG
-		LOG_MSG("Unhandled speaker mode %d",mode);
+		LOG_MSG("Unhandled speaker mode %u",static_cast<uint32_t>(mode));
 #endif
 		return;
 	}
